+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+## [2025-01-16]
+
+- Added Vuetify, Material Design Icons, and Nuxt Vuetify module dependencies
+
 ## [2025-01-15]
 
 - Disabled SSR globally to fix Vuetify form component issues